chore(onboarding): Makes onboarding route guards injectable
